pyproject.toml
[project]
name = "sales-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.4",
    "django-extensions>=4.1",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.69",
    "langchain-openai>=0.3.28",
    "langchain-postgres>=0.0.15",
    "langgraph>=0.5.3",
    "markdown>=3.8.2",
    "mypy>=1.17.0",
    "openai>=1.96.1",
    "pgvector>=0.3.6",
    "psycopg>=3.2.9",
    "psycopg-binary>=3.2.9",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.3",
    "tiktoken>=0.9.0",
]
